scripts/requirements.txt
cloudfoundry-client==1.32.0
requests==2.28.2
retry==0.9.2
python-dotenv==0.21.1
pyjwt==2.6.0
PyYAML==6.0